requirements.txt
python-telegram-bot==20.7
requests==2.31.1
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
apscheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3